util/mem: move alloc helper to util section
Move the alloc helpers to its own header file so that we avoid an
include header recursion when defining auto cleanup helpers.
