+++ util/meson.build
@@ -5,6 +5,7 @@ sources += [
   'util/base64.c',
   'util/cleanup.c',
   'util/crc32.c',
+  'util/mem.c',
   'util/suffix.c',
   'util/types.c',
 ]
